Factor out "ingredient handler utilities" into a separate module. This is a prerequisite for factoring out the handlers themselves.
This also moves the `source` variable into `context`, because `source`
needs to be accessible in both the ingredient handlers and the utils.
